Toolbox (API) Skeleton
Created a skelton for a MatLab Toolbox which we can use as an API to contain our wrappers, etc. Just run "ADD_P2TOOLBOX_TO_PATH.m", use "CHANGE_FOLDER," and it should work. Lmk if not.

If you're unfamiliar with Toolboxes, see "placeholder.m" as an example of how methods are contained in the toolbox (it's basically just a directory).